Refactor: 예외 처리 전략 변경
