+++ src/main/java/controller/IndexController.java
@@ -4,7 +4,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import support.annotation.Controller;
 import support.annotation.RequestMapping;
-import support.exception.FoundException;
+import support.web.exception.FoundException;
 import support.web.HttpMethod;
 import support.web.ModelAndView;
