+++ src/main/java/controller/PostViewController.java
@@ -8,7 +8,7 @@ import org.slf4j.LoggerFactory;
 import support.annotation.Controller;
 import support.annotation.PathVariable;
 import support.annotation.RequestMapping;
-import support.exception.FoundException;
+import support.web.exception.FoundException;
 import support.web.HttpMethod;
 import support.web.ModelAndView;
 import utils.LoginUtils;
