+++ src/main/java/controller/UserController.java
@@ -8,7 +8,7 @@ import org.slf4j.LoggerFactory;
 import support.annotation.Controller;
 import support.annotation.RequestMapping;
 import support.annotation.RequestParam;
-import support.exception.FoundException;
+import support.web.exception.FoundException;
 import support.web.HttpMethod;
 import support.web.ModelAndView;
 import utils.LoginUtils;
