--- src/main/java/exception/ExceptionName.java
@@ -1,9 +1,0 @@
-package exception;
-
-public abstract class ExceptionName {
-
-    public static final String FILE_NOT_FOUND = "해당 경로에 파일이 존재하지 않습니다.";
-    public static final String WRONG_ARGUMENT = "잘못된 인자입니다.";
-    public static final String ALREADY_REGISTERED = "이미 등록되어 있는 인스턴스입니다.";
-
-}
