--- src/main/java/support/exception/BadRequestException.java
@@ -1,11 +1,0 @@
-package support.exception;
-
-import webserver.response.HttpStatus;
-
-public class BadRequestException extends HttpException {
-
-    public BadRequestException(String message) {
-        super(message, HttpStatus.BAD_REQUEST);
-    }
-
-}
