+++ src/main/java/support/web/ViewResolver.java
@@ -3,8 +3,8 @@ package support.web;
 import model.Session;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import support.exception.NotFoundException;
-import support.exception.ServerErrorException;
+import support.web.exception.NotFoundException;
+import support.web.exception.ServerErrorException;
 import support.instance.DefaultInstanceManager;
 import support.web.view.View;
 import support.web.view.ViewFactory;
