+++ src/main/java/support/web/exception/BadRequestException.java
@@ -1,0 +1,11 @@
+package support.web.exception;
+
+import webserver.response.HttpStatus;
+
+public class BadRequestException extends HttpException {
+
+    public BadRequestException() {
+        super(HttpStatus.BAD_REQUEST);
+    }
+
+}
