+++ src/main/java/support/web/exception/NotSupportedException.java
@@ -1,4 +1,4 @@
-package support.exception;
+package support.web.exception;
 
 public class NotSupportedException extends Exception {
 }
